+++ Changelog.md
@@ -1,9 +1,10 @@
 # Project GoldStars Calculator S Changelog
 
 ## Version 3.1.x
-### Version 3.1
+### Version 3.1 (March 3, 2020)
 * Add the ability to convert to and from inches in Distance Conversions.
 * In the "Distance Conversions" window, pressing Enter in one of the input fields has the same effect as clicking the corresponding button.
+* Update the default un-maximized size of the main window.
 * Other minor changes.
 
 ## Version 3.0.x